Solve issue of not reading the second line for solid solutions
Until then, when solid solution end-members were spread on 2 lines, only the first line was read. I added a part that checks for this and read the data, store them and append the top (end-member names) and line (end-member proportions). I changed the concat to merge because when the last log zi was spread on 2 lines and the end-members weren't matching the previous log zis, pd.concat() add index issues.